+++ go/src/student_submissions/movielens/hugo_vega/file_splitting_test.go
@@ -1,4 +1,4 @@
-package fileprocessing
+package main
 
 import (
 	"bufio"
